preview-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
aio-pika==9.4.3
pytest==7.4.3
pytest-mock==3.12.0
httpx==0.25.2
